Limit regex on wp database tables leaking
Current command line on the database used a regex to target tables based
on blog id. This however had a bug there the regex wildcard couldn't
determine between 2 and 29 ... anything that started with a 2 would be
pulled in. This fix limits the init to only that and wildcard on the
number is removed.